requirements.txt
python-dotenv       # para manejo de variables de entorno
groq==0.13.1        # SDK de Groq para usar modelos (si optas por Groq)
